Copy 'res' folder to device for run-tests
Sometimes tests need to push resources to target for a run-test.
The scripts currently do not have an option how to bundle additional
files with the test. To this end, this patch tests for presence of
folder named 'res' and if it exists, copies it into DEX_LOCATION on
the device.

Test 071 was converted to become the first test using this new feature.

Test: art/test.py -b -r -t 071-dexfile-get-static
Change-Id: I7eac4b15ca19fce6c616f5cc732b7ac65b8174b8