Reduce flakiness of 004-NativeAllocations test.
Under a GSS configuration, calls to registerNativeAllocation trigger
blocking GC instead of concurrent GC, which means the main thread in the
test can end up holding on to deadlockLock for longer than the finalizer
timeout.

The fix is to reduce the calls to registerNativeAlloation required to
trigger the deadlock scenario.

Bug: 68792448
Test: ./test/testrunner/testrunner.py -b -t 004-NativeAllocations
Test: Verify test catches deadlock case on buggy versions of art.
Change-Id: Ide417de7a5cad593afd2877054f88588b7a32be3