Enable profile data filtering in profman
Update profile merging to accept a set of apks (passes with --apk) which
will dictate what data should be processed.

When profman is invoked with a list of --apk files, only profile data
belonging to that apks will be in the output reference profile.

If no --dex-location is specified then the locations is inferred from
reding /proc/self/fd/apk_fd link.

Test: profile_assistant_test
Bug: 30934496
Change-Id: I44698c6db545ecf91454db1387c3d0e47fe5b9b3