Make exception test more robust against dex changes.
Rationale:
With the shift from dx to d8, dex bytecode may be layed out
slightly different. The exception test was not particularly
robust against this, as it was looking for instruction at
dex_pc == 3 in method g(). Incidentialy, the way it was setup,
it would even use that dex_pc for querying f(), which seemed
unintentional (and makes it hard to make the test more robust
by e.g. looking up the new instance instruction). Therefore,
test was made more robust by simply assuming dex_pc == 0,
which consistently maps to first statement in source.

Bug: 73050355

Test: exception_test
Change-Id: Ic5b2784a48ae10d139c0b77a3bbfcc53b58a5b42