Refactor hidden API runtime flag from negative to positive
There are only two situations in which we want to enable hidden API
access checks: (a) when forking a non-platform app from zygote, and
(b) when compiling such app with dex2oat. It is easier to cover these
two situations than all the other situations when we want to disable
the checks. Make the flag in Runtime class false by default and only
enable if checks are explicitly requested by the runtime.

Bug: 64382372
Test: boot device, install and run apps, check log messages
Change-Id: I7faf440ed714a5ddd08f6172d370202f0131db50