Ensure redefinition does not break breakpoint deoptimization
We had a bug where redefining a method with breakpoints made it
impossible to deoptimize the (newly redefined) method. This meant that
breakpoints added after the redefinition might not work. We fixed this
by removing the breakpoints earlier in the redefinition process,
when we were still capable of removing the deoptimization.

Bug: 73742185
Test: ./test.py --host -j50
Change-Id: I86973fbdce370a3ae7bc78b5fc71470d99046044