Filter out ddms error message in 1940 better
Since the ddms error message in test 1940 is printed from within the
runtime it lacks the synchronization with System.out that other prints
have. This could cause the output to get mixed slightly with other
messages from the test. In order to fix this we filter it out using
python code that is able to understand the interleaving.

Test: ./test.py --host -j50
Change-Id: Ic42816beeea2f04046b5bb88e17a1f6bde19a58e