Move jit startup attempt to openjdkjvmti from adbconnection
We were attempting to start the jit in adbconnection if it was not
already running in order to improve the performance of debugging
system_server and other similar processes. Since we have opened up the
kArtTiVersion (0x70010200) environment to any agent (on userdebug/eng
builds) it would be useful for these other agents also have the
performance boost when running on these processes.

Test: Build
Test: adb root &&
      adb shell setenforce 0 &&
      adb forward tcp:12345 jdwp:`adb shell pidof system_server` &&
      jdb -attach localhost:12345;
      <Do things with debugger and device>

Bug: 78119634
Bug: 78195998
Change-Id: Ibfe10b43dcd23d615c03c12af2625e2b0c54f6a4