Set hidden API flags of intrinsics
Intrinsics overwrite the hidden API access flags of the respective Java
method, which is why we need to hardcode their API list membership.

VarHandle intrinsics are blacklisted because they could be used to
bypass hidden API checks and given they are new in P, should not be
used by anybody except tests (that do not enforce API checks).

The remaining intrinsics which happen to be @hide are put on light
greylist. We used to put them on whitelist implicitly, hence never
saw warnings about them. To be safe, we put them on light grey.

Note that these are set even for the core image that currently does
not have hidden API flags. That is fine because (a) VarHandles can
still be tested, and (b) light greylist membership may print warnings
but will not change the semantics.

Bug: 64382372
Bug: 77733081
Test: make
Change-Id: Ia9a7765260acb533560676e7dfcd51065cfb247d