AOT inlined method lookup should stay within dex file.
Rewrite the AOT inlined method lookup and guard against
crossing dex file boundary. The compiler does not currently
support inlining across dex files, so this is an indication
of multiple dex files defining the same class and the AOT
code having used a definition which is not used at runtime.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing --jit
Test: Pixel 2 XL boots.
Test: testrunner.py --target --optimizing --jit
Bug: 74410240
Change-Id: Ic2a595ee798e7973156bf29d1a1bcc8615405407